fix: skip admin info in setup wizard
