+++ package.json
@@ -189,7 +189,7 @@
 		"@rocket.chat/memo": "^0.31.3",
 		"@rocket.chat/message-parser": "^0.31.3",
 		"@rocket.chat/mp3-encoder": "^0.24.0",
-		"@rocket.chat/onboarding-ui": "^0.31.3",
+		"@rocket.chat/onboarding-ui": "^0.6.3-dev.431",
 		"@rocket.chat/string-helpers": "^0.31.3",
 		"@rocket.chat/ui-kit": "^0.31.3",
 		"@slack/client": "^4.12.0",
